Manage states scripts rewritted, close issue #65

+++ src/odcommon/usr/local/opendomo/services/syscript/delSystemState.sh
@@ -6,20 +6,19 @@
 # Copyright(c) 2011 OpenDomo Services SL. Licensed under GPL v3 or later
 
 STATESDIR=/etc/opendomo/states
+TEMPSTATE="/var/opendomo/tmp/state.tmp"
 STATE=$1
 
-if test -z $STATE; then
-	echo "#> Delete states"
-	echo "#ERR No system state specified"
-	exit 1
+if ! test -z $STATE; then
+    test $STATE != active && rm -r $STATESDIR/$STATE
+elif test -f $TEMPSTATE; then
+    STATE=`cat $TEMPSTATE`
+    test $STATE != active && rm -r $STATESDIR/$STATE
 else
-	if [ "$STATE" = "active" ]; then
-		echo "#ERR State 'active' can't be deleted"
-	else
-		echo "#INFO State $STATE deleted"
-		rm -r $STATESDIR/$STATE
-	fi
-
-	# Return to manage states
-	/usr/local/opendomo/manageSystemStates.sh
+    echo "#> Delete states"
+    echo "#ERR No system state specified"
+    exit 1
 fi
+
+# Return to manage states
+/usr/local/opendomo/manageSystemStates.sh
